fix(app): disambiguate cos() overload in ScanningView radial sweep
CI's Swift toolchain flags `cos(beamRad)` as an ambiguous call between
CoreGraphics.cos(CGFloat) and _math.cos(Double) once mixed with the
surrounding CGFloat arithmetic. Declaring beamRad as CGFloat up front
removes the ambiguity so the release build succeeds on every toolchain.

+++ app/Sources/SpaceMeshApp/ScanningView.swift
@@ -145,7 +145,7 @@ struct ScanningView: View {
     private func drawSweep(
         ctx: inout GraphicsContext, center: CGPoint, radius: CGFloat, beamAngle: Double
     ) {
-        let beamRad = Angle(degrees: beamAngle).radians
+        let beamRad = CGFloat(Angle(degrees: beamAngle).radians)
         // 잔광 부채꼴 (여러 겹으로 페이드)
         let trailSteps = 24
         let trailSpan = 70.0
